Updating win-release.sh

+++ win-release.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
+# This is script is meant to be run from a MSYS console
+
 MINGW_PATH=C:/MinGW/bin/
 OUTPUT=kissplayer
+VERSION=`grep AC_INIT configure.ac | cut -d',' -f2 | tr -d '[] '`
 
 containsElement() {
   local e
@@ -16,20 +19,24 @@ mkdir -p $OUTPUT
 cd po
 
 find . -type f -iname "*.gmo" -print0 | while IFS= read -r -d $'\0' file; do
-    lang=`echo "$file" | cut -d'.' -f2 |cut -d'/' -f2`
-    mkdir -p ../$OUTPUT/locale/$lang/LC_MESSAGES
-    cp "$file" ../$OUTPUT/locale/$lang/LC_MESSAGES/kissplayer.mo
+  lang=`echo "$file" | cut -d'.' -f2 | tr -d '/'`
+  mkdir -p ../$OUTPUT/locale/$lang/LC_MESSAGES
+  cp "$file" ../$OUTPUT/locale/$lang/LC_MESSAGES/kissplayer.mo
 done
 
 cd ..
 
 objdump -p kissplayer.exe | grep "DLL Name:" | while read -r dll; do
-    dll=`echo $dll | cut -d' ' -f3`
-    containsElement $dll "${win_dlls[@]}"
-    if [ "$?" -eq "1" ]; then
-        continue
-    fi
-    cp $MINGW_PATH$dll $OUTPUT
+  dll=`echo $dll | cut -d' ' -f3`
+  containsElement $dll "${win_dlls[@]}"
+  if [ "$?" -eq "1" ]; then
+    continue
+  fi
+  if [ ! -f $MINGW_PATH$dll ]; then
+    echo "$MINGW_PATH$dll not found!"
+    continue
+  fi
+  cp $MINGW_PATH$dll $OUTPUT
 done
 
 strip kissplayer.exe
@@ -38,4 +45,6 @@ cp LICENSE.txt $OUTPUT
 cp CHANGELOG.txt $OUTPUT
 cp README.txt $OUTPUT
 
-7z a -tzip kissplayer kissplayer
+7z a -tzip "kissplayer-$VERSION.zip" kissplayer > /dev/null
+
+echo done
